oxigraph

Module sparql

SPARQL implementation.

The entry point for SPARQL execution is the SparqlEvaluator type.

Modules

results

Utilities to read and write RDF results formats using sparesults.

Structs

BoundPreparedSparqlQuery

A prepared SPARQL query bound to a storage, ready to be executed.

BoundPreparedSparqlUpdate

A prepared SPARQL query bound to a storage, ready to be executed.

CancellationToken

A token that can be used to mark something as canceled.

PreparedSparqlQuery

A prepared SPARQL query.

PreparedSparqlUpdate

A prepared SPARQL update.

QueryDeprecated

A parsed SPARQL query.

QueryDatasetSpecification

An extended SPARQL query dataset specification.

QueryExplanation

The explanation of a query.

QuerySolution

Tuple associating variables and terms that are the result of a SPARQL query.

QuerySolutionIter

An iterator over QuerySolutions.

QueryTripleIter

An iterator over the triples that compose a graph solution.

SparqlEvaluator

SPARQL evaluator.

SparqlSyntaxError

Error returned during SPARQL parsing.

UpdateDeprecated

A parsed SPARQL update.

Variable

A SPARQL query owned variable.

VariableNameParseError

An error raised during Variable name validation.

Enums

QueryEvaluationError

A SPARQL evaluation error

QueryResults

Results of a SPARQL query.

UpdateEvaluationError

An error from SPARQL UPDATE evaluation

Traits

AggregateFunctionAccumulator

A trait for custom aggregate function implementation.

DefaultServiceHandler

Default handler for SPARQL 1.1 Federated Query SERVICEs.

ServiceHandler

Handler for SPARQL 1.1 Federated Query SERVICEs.

Type Aliases

EvaluationErrorDeprecated

QueryDataset

QueryOptionsDeprecated

UpdateOptionsDeprecated