oxigraph::sparql

Enum EvaluationError

#[non_exhaustive]
pub enum EvaluationError {

    Parsing(SparqlSyntaxError),
    Storage(StorageError),
    GraphParsing(RdfParseError),
    ResultsParsing(QueryResultsParseError),
    ResultsSerialization(Error),
    Service(Box<dyn Error + Send + Sync + 'static>),
    GraphAlreadyExists(NamedNode),
    GraphDoesNotExist(NamedNode),
    UnboundService,
    UnsupportedService(NamedNode),
    UnsupportedContentType(String),
    ServiceDoesNotReturnSolutions,
    NotAGraph,

}

A SPARQL evaluation error.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Parsing(SparqlSyntaxError)

An error in SPARQL parsing.

Storage(StorageError)

An error from the storage.

GraphParsing(RdfParseError)

An error while parsing an external RDF file.

ResultsParsing(QueryResultsParseError)

An error while parsing an external result file (likely from a federated query).

ResultsSerialization(Error)

An error returned during results serialization.

Service(Box<dyn Error + Send + Sync + 'static>)

Error during SERVICE evaluation

GraphAlreadyExists(NamedNode)

Error when CREATE tries to create an already existing graph

GraphDoesNotExist(NamedNode)

Error when DROP or CLEAR tries to remove a not existing graph

UnboundService

The variable storing the SERVICE name is unbound

UnsupportedService(NamedNode)

The given SERVICE is not supported

UnsupportedContentType(String)

The given content media type returned from an HTTP response is not supported (SERVICE and LOAD)

ServiceDoesNotReturnSolutions

The SERVICE call has not returns solutions

NotAGraph

The results are not a RDF graph