oxigraph::sparql

Type Alias EvaluationError

pub type EvaluationError = QueryEvaluationError;

👎Deprecated since 0.5.0: Use QueryEvaluationError instead

Aliased Type

pub enum EvaluationError {
    Dataset(Box<dyn Error + Send + Sync>),
    Service(Box<dyn Error + Send + Sync>),
    NotExistingSubstitutedVariable(Variable),
    UnexpectedDefaultGraph,
    UnsupportedCustomFunction(NamedNode),
    UnsupportedCustomFunctionArity {
        name: NamedNode,
        expected: RangeInclusive<usize>,
        actual: usize,
    },
    UnboundService,
    InvalidServiceName(Term),
    UnsupportedService(NamedNode),
    InvalidStorageTripleTerm,
    Cancelled,
}

Variants

Dataset(Box<dyn Error + Send + Sync>)

Error from the underlying RDF dataset

Service(Box<dyn Error + Send + Sync>)

Error during SERVICE evaluation

NotExistingSubstitutedVariable(Variable)

If a variable present in the given initial substitution is not present in the SELECT part of the query

UnexpectedDefaultGraph

Error if the dataset returns the default graph even if a named graph is expected

UnsupportedCustomFunction(NamedNode)

The given custom function is not supported

UnsupportedCustomFunctionArity

The given custom function arity is not supported

Fields

name: NamedNode

expected: RangeInclusive<usize>

actual: usize

UnboundService

The variable storing the SERVICE name is unbound

InvalidServiceName(Term)

Invalid service name

UnsupportedService(NamedNode)

The given SERVICE is not supported

InvalidStorageTripleTerm

Cancelled