oxigraph
Module io

Utilities to read and write RDF graphs and datasets using OxRDF I/O.

The entry points of this module are the two RdfParser and RdfSerializer structs.

Usage example converting a Turtle file to a N-Triples file:

use oxigraph::io::{RdfFormat, RdfParser, RdfSerializer};

let turtle_file = b"@base <http://example.com/> .
@prefix schema: <http://schema.org/> .
<foo> a schema:Person ;
    schema:name \"Foo\" .
<bar> a schema:Person ;
    schema:name \"Bar\" .";

let ntriples_file = b"<http://example.com/foo> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://schema.org/Person> .
<http://example.com/foo> <http://schema.org/name> \"Foo\" .
<http://example.com/bar> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://schema.org/Person> .
<http://example.com/bar> <http://schema.org/name> \"Bar\" .
";

let mut serializer = RdfSerializer::from_format(RdfFormat::NTriples).for_writer(Vec::new());
for quad in RdfParser::from_format(RdfFormat::Turtle).for_reader(turtle_file.as_ref()) {
    serializer.serialize_quad(&quad.unwrap()).unwrap();
}
assert_eq!(serializer.finish().unwrap(), ntriples_file);

Re-exports

pub use self::read::DatasetParser;

Deprecated

pub use self::read::GraphParser;

Deprecated

pub use self::write::DatasetSerializer;

Deprecated

pub use self::write::GraphSerializer;

Deprecated

Modules

read

Utilities to read RDF graphs and datasets.

write

Utilities to write RDF graphs and datasets.

Structs

RdfParser

Parsers for RDF serialization formats.

RdfSerializer

A serializer for RDF serialization formats.

RdfSyntaxError

An error in the syntax of the parsed file.

ReaderQuadParser

Parses a RDF file from a Read implementation.

SliceQuadParser

Parses a RDF file from a byte slice.

TextPosition

A position in a text i.e. a line number starting from 0, a column number starting from 0 (in number of code points) and a global file offset starting from 0 (in number of bytes).

WriterQuadSerializer

Serializes quads or triples to a Write implementation.

Enums

DatasetFormatDeprecated

RDF dataset serialization formats.

GraphFormatDeprecated

RDF graph serialization formats.

RdfFormat

RDF serialization formats.

RdfParseError

Error returned during RDF format parsing.