oxigraph::io

Enum RdfFormat

#[non_exhaustive]
pub enum RdfFormat {
    N3,
    NQuads,
    NTriples,
    RdfXml,
    TriG,
    Turtle,
    JsonLd {
        profile: JsonLdProfileSet,
    },
}

RDF serialization formats.

This enumeration is non exhaustive. New formats might be added in the future.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

N3

N3

NQuads

N-Quads

NTriples

N-Triples

RdfXml

RDF/XML

TriG

TriG

Turtle

Turtle

JsonLd

JSON-LD

Fields

profile: JsonLdProfileSet

Implementations

impl RdfFormat

pub const fn iri(self) -> &'static str

The format canonical IRI according to the Unique URIs for file formats registry.

use oxrdfio::RdfFormat;

assert_eq!(
    RdfFormat::NTriples.iri(),
    "http://www.w3.org/ns/formats/N-Triples"
)

pub const fn media_type(self) -> &'static str

The format IANA media type.

use oxrdfio::RdfFormat;

assert_eq!(RdfFormat::NTriples.media_type(), "application/n-triples")

pub const fn file_extension(self) -> &'static str

The format IANA-registered file extension.

use oxrdfio::RdfFormat;

assert_eq!(RdfFormat::NTriples.file_extension(), "nt")

pub const fn name(self) -> &'static str

The format name.

use oxrdfio::RdfFormat;

assert_eq!(RdfFormat::NTriples.name(), "N-Triples")

pub const fn supports_datasets(self) -> bool

Checks if the formats supports RDF datasets and not only RDF graphs.

use oxrdfio::RdfFormat;

assert_eq!(RdfFormat::NTriples.supports_datasets(), false);
assert_eq!(RdfFormat::NQuads.supports_datasets(), true);

pub const fn supports_rdf_star(self) -> bool

Checks if the formats supports RDF-star quoted triples.

use oxrdfio::RdfFormat;

assert_eq!(RdfFormat::NTriples.supports_rdf_star(), true);
assert_eq!(RdfFormat::RdfXml.supports_rdf_star(), false);

pub fn from_media_type(media_type: &str) -> Option<RdfFormat>

Looks for a known format from a media type.

It supports some media type aliases. For example, โapplication/xmlโ is going to return RdfFormat::RdfXml even if it is not its canonical media type.

Example:

use oxjsonld::JsonLdProfile;
use oxrdfio::RdfFormat;

assert_eq!(
    RdfFormat::from_media_type("text/turtle; charset=utf-8"),
    Some(RdfFormat::Turtle)
);
assert_eq!(
    RdfFormat::from_media_type(
        "application/ld+json ; profile = http://www.w3.org/ns/json-ld#streaming"
    ),
    Some(RdfFormat::JsonLd {
        profile: JsonLdProfile::Streaming.into()
    })
)

pub fn from_extension(extension: &str) -> Option<RdfFormat>

Looks for a known format from an extension.

It supports some aliases.

Example:

use oxrdfio::RdfFormat;

assert_eq!(RdfFormat::from_extension("nt"), Some(RdfFormat::NTriples))