oxigraph
Module sparql

SPARQL implementation.

Stores execute SPARQL. See Store for an example.

Modules

results

Utilities to read and write RDF results formats using sparesults.

Structs

Query

A parsed SPARQL query.

QueryDataset

A SPARQL query dataset specification

QueryExplanation

The explanation of a query.

QueryOptions

Options for SPARQL query evaluation.

QuerySolution

Tuple associating variables and terms that are the result of a SPARQL query.

QuerySolutionIter

An iterator over QuerySolutions.

QueryTripleIter

An iterator over the triples that compose a graph solution.

SparqlSyntaxError

Error returned during SPARQL parsing.

Update

A parsed SPARQL update.

UpdateOptions

Options for SPARQL update evaluation.

Variable

A SPARQL query owned variable.

VariableNameParseError

An error raised during Variable name validation.

Enums

EvaluationError

A SPARQL evaluation error.

QueryResults

Results of a SPARQL query.

Traits

ServiceHandler

Handler for SPARQL 1.1 Federated Query SERVICE.